exercise 5.4

+++ part4/bloglist/controllers/blogs.js
@@ -1,18 +1,14 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 
-
 blogsRouter.get('', async (request, response) => {
-  Blog.find({}).then(blogs => {
-    blogs = blogs.map(blog => {
-      return { id: blog._id.toString(), title: blog.title, author: blog.author, likes: blog.likes }
-    })
+  Blog.find({}).populate('user', { blogs: 0 }).then(blogs => {
     response.json(blogs)
   })
 })
 
 blogsRouter.get('/:id', async (request, response, next) => {
-  Blog.findById(request.params.id).then(blog => {
+  Blog.findById(request.params.id).populate('user', { blogs: 0 }).then(blog => {
     if (!blog) {
       response.send(400).end()
     } else {
@@ -22,6 +18,11 @@ blogsRouter.get('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!request.user || blog?.user?.toString() !== request.user.id.toString()) {
+    return response.status(401).json({ error: 'only creator of the blog can delete' })
+  }
+
   Blog.findByIdAndRemove(request.params.id).then(result => {
     if (result === null) {
       response.status(400).end()
@@ -32,20 +33,35 @@ blogsRouter.delete('/:id', async (request, response, next) => {
 })
 
 blogsRouter.post('', async (request, response, next) => {
-  const blog = new Blog(request.body)
-  blog.likes = blog.likes || 0
-  if (!blog.title || !blog.url) {
-    response.status(400)
-  }
+  const user = request.user
+
+  const blog = new Blog({
+    title: request.body.title,
+    url: request.body.url,
+    author: request.body.author,
+    user: user._id
+  })
+
+  blog.likes = request.body.likes || 0
   try {
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     response.status(201).json(savedBlog)
-  } catch (e) {
-    next(e)
+  } catch (error) {
+    console.log(error)
+    response.status(400).json(error)
+    next(error)
   }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if (!request.user || blog?.user?.toString() !== request.user.id.toString()) {
+    return response.status(401).json({ error: 'only creator of the blog can update' })
+  }
+
   Blog.findByIdAndUpdate(request.params.id, request.body, {
     new: true, runValidators: true, context: 'query'
   }).then(blog => {
